push

+++ app/Http/Requests/VenueRequest.php
@@ -31,10 +31,10 @@ class VenueRequest extends FormRequest
                 'required', 'min:3'
             ],
             'longitude' => [
-                'required', 'integer', 'between:-180,180'
+                'required', 'numeric', 'between:-180,180'
             ],
             'latitude' => [
-                'required','integer','between:-90,90'
+                'required','numeric','between:-90,90'
             ]
         ];
     }
